Add of a confirm on plugin removal

+++ vimconfig.sh
@@ -192,9 +192,11 @@ dotvim_remove() {
   plugin_name="$(to_bundle_name "$1")"
 
   if [ -d "bundle/$plugin_name" ]; then
-    git submodule deinit -f "bundle/$plugin_name"
-    git rm "bundle/$plugin_name"
-    git commit -m "Remove $plugin_name bundle from submodules list."
+    if confirm "Remove plugin $plugin_name?"; then
+      git submodule deinit -f "bundle/$plugin_name"
+      git rm "bundle/$plugin_name"
+      git commit -m "Remove $plugin_name plugin from submodules list."
+    fi
   else
     echo "The $plugin_name plugin does not exist under bundle/. Doing nothing..."
     return 2
